Showing running pipelines + grouped history-related functions in History.kt
